{actions,behavior}.py: move Session into behavior schema:
